k2tf: Avoid adding a tf.identity tensor at the graph outputs
Removed a code that generate a tf.identity tensor for each of the model's
outputs. This identity tensor op added a strided-slice op with non zero
shrink_axis_mask which prevent tensor-rt from optimizing the graph.
This might cause some difficulties in accessing the model outputs with TF (as
was state on a code remark: 'Alias the outputs in the model - this sometimes
makes them easier to access in TF').

Signed-off-by: ep <ep@no.mail>